requirements.txt
pandas==2.2.0
streamlit==1.30.0
plotly.express==0.4.1
altair==5.2.0
